Migrate service to jooq

+++ server/src/main/java/net/serlith/version/server/types/SubmitBuildRequest.java
@@ -3,6 +3,6 @@ package net.serlith.version.server.types;
 public record SubmitBuildRequest(
         String software,
         String version,
-        int build
+        long build
 ) {
 }
